refactor: Reduce redundant page-title hooks
